added retrofit implementation of the sephora api

+++ app/src/main/java/com/hilfritz/android/viper/data/sephoraApi/pojo/SephoraProductRepositoryImpl.java
@@ -19,7 +19,7 @@ public class SephoraProductRepositoryImpl implements SephoraProductRepository{
     }
 
     @Override
-    public Observable<ProductsListWrapper> getProductsByPageAndCategory(long page, String cateogry) {
+    public Observable<ProductsListWrapper> getProductsByPageAndCategory(int page, String cateogry) {
         return null;
     }
